Rename TagDataIterator to cache_files
