+++ xlive/Blam/Engine/cache/cache_files.cpp
@@ -1,6 +1,6 @@
 #include "stdafx.h"
 
-#include "TagDataIterator.h"
+#include "cache_files.h"
 
 datum s_tag_data_iterator::get_next_datum()
 {
